src/requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
PyYAML
IPy
requests
gevent
gunicorn